+++ test/wasm2wat/not_exists.t
@@ -1,0 +1,5 @@
+  $ owi wasm2wat idontexist.wat
+  owi: no file 'idontexist.wat'
+  Usage: owi wasm2wat [OPTION]… [ARG]…
+  Try 'owi wasm2wat --help' or 'owi --help' for more information.
+  [124]