+++ CHANGES.md
@@ -1,11 +1,10 @@
 ## unreleased
 
 - use a subcommand system for the `owi` binary
-- add `owi c`, `owi fmt`, `owi opt`, `owi sym` and `owi validate` subcommands
+- add `owi c`, `owi fmt`, `owi opt`, `owi sym`, `owi validate` and `owi wasm2wat` subcommands
 - add a fuzzer
 - add a profiling mode
 - make `spectec` print stuff for real
-- add a binary parser
 
 ## 0.1 - 2023-01-14